+++ src/backend/src/main/java/com/personal/finance/backend/reports/dto/response/DashboardOverviewDTO.java
@@ -3,20 +3,20 @@ package com.personal.finance.backend.reports.dto.response;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
 @Setter
 public class DashboardOverviewDTO {
-    private Double totalBalance;
-    private Double totalIncome;
-    private Double totalExpense;
-    private Double netSavings;
+    private BigDecimal totalBalance;
+    private BigDecimal totalIncome;
+    private BigDecimal totalExpense;
+    private BigDecimal netSavings;
 
     private Double incomeChangePercent;
     private Double expenseChangePercent;
 
     private List<CategoryExpenseDTO> expenseByCategory;
-
     private List<TrendDataDTO> trendData;
 }
